Fix accidental duplication of "Homepage" entry in the README. This was my bad
While editing the README.md, I had accidently hit 'p' on my keyboard
twice. Because I was using a vim-like text editor, this pasted whatever I
had copied (or yanked). It turns out what I had copied was the Homepage
entry. I did this so I wouldn't have to retype the line myself, using
the entry as a sort of template, if you will. This commit removes that
duplicate entry.